fix pkg
