+++ app/src/main/java/com/example/androidtechsample/ui/PlaygroundFragment.kt
@@ -1,4 +1,4 @@
-package com.example.androidtechsample
+package com.example.androidtechsample.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
+import com.example.androidtechsample.R
 import com.example.androidtechsample.databinding.FragmentPlaygroundBinding
 import dagger.hilt.android.AndroidEntryPoint
