+++ app/src/main/java/com/example/androidtechsample/ui/motion/MotionCustomFragment.kt
@@ -1,22 +1,22 @@
-package com.example.androidtechsample.motion
+package com.example.androidtechsample.ui.motion
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.androidtechsample.databinding.FragmentMotionKeyPositionBinding
+import com.example.androidtechsample.databinding.FragmentMotionCustomBinding
 
-class MotionKeyPositionFragment : Fragment() {
+class MotionCustomFragment: Fragment() {
 
-  private lateinit var binding: FragmentMotionKeyPositionBinding
+  private lateinit var binding: FragmentMotionCustomBinding
 
   override fun onCreateView(
     inflater: LayoutInflater,
     container: ViewGroup?,
     savedInstanceState: Bundle?
   ): View {
-    binding = FragmentMotionKeyPositionBinding.inflate(inflater)
+    binding = FragmentMotionCustomBinding.inflate(inflater)
     return binding.root
   }
 }
