+++ app/src/main/java/com/example/androidtechsample/ui/motion/MotionKeyPositionFragment.kt
@@ -1,22 +1,22 @@
-package com.example.androidtechsample.motion
+package com.example.androidtechsample.ui.motion
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.androidtechsample.databinding.FragmentMotionCustomBinding
+import com.example.androidtechsample.databinding.FragmentMotionKeyPositionBinding
 
-class MotionCustomFragment: Fragment() {
+class MotionKeyPositionFragment : Fragment() {
 
-  private lateinit var binding: FragmentMotionCustomBinding
+  private lateinit var binding: FragmentMotionKeyPositionBinding
 
   override fun onCreateView(
     inflater: LayoutInflater,
     container: ViewGroup?,
     savedInstanceState: Bundle?
   ): View {
-    binding = FragmentMotionCustomBinding.inflate(inflater)
+    binding = FragmentMotionKeyPositionBinding.inflate(inflater)
     return binding.root
   }
 }
