+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".MainActivity">
+  tools:context=".ui.MainActivity">
 
   <androidx.fragment.app.FragmentContainerView
     android:id="@+id/nav_host_fragment"
